// src/app/model/kimarite.ts
enum KIMARITE {
  YORIKIRI = '&#23492;&#12426;&#20999;&#12426;',
  YORITAOSHI = '&#23492;&#12426;&#20498;&#12375;',
  OSHIDASHI = '&#25276;&#12375;&#20986;&#12375;',
  OSHITAOSHI = '&#25276;&#12375;&#20498;&#12375;',
  HATAKIKOMI = '&#12399;&#12383;&#12365;&#36796;&#12415;',
  TSUKIDASHI = '&#31361;&#12365;&#20986;&#12375;',
  OKURIDASHI = '&#36865;&#12426;&#20986;&#12375;',
  TSUKIOTOSHI = '&#31361;&#12365;&#33853;&#12375;',
  HIKIOTOSHI = '&#24341;&#33853;&#12375;',
  UWATENAGE = '&#19978;&#25163;&#25237;&#12370;',
  SUKUINAGE = '&#12377;&#12367;&#12356;&#25237;&#12370;',
  SHITATENAGE = '&#19979;&#25163;&#25237;&#12370;',
  UWATEDASHINAGE = '&#19978;&#25163;&#20986;&#12375;&#25237;&#12370;',
  KOTENAGE = '&#23567;&#25163;&#25237;&#12370;',
  KATASUKASHI = '&#32937;&#12377;&#12363;&#12375;',
  KIMEDASHI = '&#12365;&#12417;&#12384;&#12375;',
  SHITATEDASHINAGE = '&#19979;&#25163;&#20986;&#12375;&#25237;&#12370;',
  TOTTARI = '&#12392;&#12387;&#12383;&#12426;',
  TSUKITAOSHI = '&#31361;&#12365;&#20498;&#12375;',
  KUBINAGE = '&#39318;&#25237;&#12370;',
  WATASHIKOMI = '&#31169;&#36796;&#12415;',
  HIKKAKE = '&#12402;&#12363;&#12387;&#12369;',
  OKURITAOSHI = '&#36865;&#12426;&#20498;&#12375;',
  SOTOGAKE = '&#22806;&#25499;&#12369;',
  TSURIDASHI = '&#37347;&#12426;&#20986;&#12375;',
  SUSOHARAI = '&#12377;&#12381;&#25173;&#12356;',
  KIRIKAESHI = '&#20999;&#12426;&#36820;&#12375;',
  SHITATEHINERI = '&#19979;&#25163;&#12402;&#12397;&#12426;',
  ABISETAOSHI = '&#28020;&#12403;&#12379;&#20498;&#12375;',
  ASHITORI = '&#36275;&#21462;&#12426;',
  UTCHARI = '&#25171;&#12387;&#26820;&#12426;',
  UWATEHINERI = '&#19978;&#25163;&#12402;&#12397;&#12426;',
  HANSOKU = '&#12399;&#12435;&#12381;&#12367;',
  ISAMIASHI = '&#21191;&#12415;&#36275;',
  KIMETAOSHI = '&#12365;&#12417;&#20498;&#12375;',
  FUSENSHO = '&#19981;&#25126;&#21213;',
}

export default KIMARITE;